requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Pillow==10.1.0
scikit-image==0.22.0
tensorflow==2.15.0
numpy==1.26.2
gunicorn==21.2.0
